+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - silo-core: add reentrancy protection for liquidation call
 - silo-vaults: fix max redeem
+- silo-vaults: ensure that deposit not generate zero shares
 
 ## [2.0.1] - 2025-03-19
 ### Fixed